+++ src/netspresso_trainer/pipelines/detection.py
@@ -1,18 +1,17 @@
-import logging
 from typing import Literal
 
 import numpy as np
 import torch
+from loguru import logger
 
 from .base import BasePipeline
 
-logger = logging.getLogger(__name__)
-
 
 class DetectionPipeline(BasePipeline):
-    def __init__(self, conf, task, model_name, model, devices, train_dataloader, eval_dataloader, class_map, **kwargs):
+    def __init__(self, conf, task, model_name, model, devices,
+                 train_dataloader, eval_dataloader, class_map, logging_dir, **kwargs):
         super(DetectionPipeline, self).__init__(conf, task, model_name, model, devices,
-                                                train_dataloader, eval_dataloader, class_map, **kwargs)
+                                                train_dataloader, eval_dataloader, class_map, logging_dir, **kwargs)
         self.num_classes = train_dataloader.dataset.num_classes
 
     def train_step(self, batch):