+++ src/netspresso_trainer/trainer_common.py
@@ -2,6 +2,7 @@ from pathlib import Path
 from typing import Literal
 
 import torch
+import torch.distributed as dist
 from omegaconf import DictConfig
 from torch.nn.parallel import DistributedDataParallel as DDP
 
@@ -9,7 +10,7 @@ from .dataloaders import build_dataloader, build_dataset
 from .models import SUPPORTING_TASK_LIST, build_model, is_single_task_model
 from .pipelines import build_pipeline
 from .utils.environment import set_device
-from .utils.logger import set_logger
+from .utils.logger import add_file_handler, get_logging_dir, set_logger
 
 
 def train_common(conf: DictConfig, log_level: Literal['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'] = 'INFO'):
@@ -36,12 +37,17 @@ def train_common(conf: DictConfig, log_level: Literal['DEBUG', 'INFO', 'WARNING'
     if is_graphmodule_training:
         model_name += "_graphmodule"
 
-    logger.info(f"Task: {task} | Model: {model_name} | Training with torch.fx model? {is_graphmodule_training}")
+    logging_dir: Path = get_logging_dir(task, model_name, output_root_dir="./outputs", distributed=distributed)
+    add_file_handler(logging_dir / "result.log", distributed=conf.distributed)
+
+    if not distributed or dist.get_rank() == 0:
+        logger.info(f"Task: {task} | Model: {model_name} | Training with torch.fx model? {is_graphmodule_training}")
+        logger.info(f"Result will be saved at {logging_dir}")
 
     if conf.distributed and conf.rank != 0:
         torch.distributed.barrier()  # wait for rank 0 to download dataset
 
-    train_dataset, valid_dataset, test_dataset = build_dataset(conf.data, conf.augmentation, task, model_name)
+    train_dataset, valid_dataset, test_dataset = build_dataset(conf.data, conf.augmentation, task, model_name, distributed=distributed)
 
     if conf.distributed and conf.rank == 0:
         torch.distributed.barrier()
@@ -63,6 +69,7 @@ def train_common(conf: DictConfig, log_level: Literal['DEBUG', 'INFO', 'WARNING'
     trainer = build_pipeline(conf, task, model_name, model,
                              devices, train_dataloader, eval_dataloader,
                              class_map=train_dataset.class_map,
+                             logging_dir=logging_dir,
                              is_graphmodule_training=is_graphmodule_training)
 
     trainer.set_train()