+++ src/netspresso_trainer/utils/logger.py
@@ -1,63 +1,58 @@
-import logging
+# import logging
+import sys
 import time
-from typing import Literal
+from pathlib import Path
+from typing import Literal, Optional, Union
 
+import torch
 import torch.distributed as dist
+from loguru import logger
 from omegaconf import DictConfig, ListConfig, OmegaConf
 
 __all__ = ['set_logger', 'yaml_for_logging']
-ROOT_LOGGER_NAME = "netspresso_trainer"
 
-class RankFilter(logging.Filter):
-    def filter(self, record):
-        try:
-            return dist.get_rank() == 0
-        except RuntimeError:  # Default process group has not been initialized, please make sure to call init_process_group.
-            return True
+OUTPUT_ROOT_DIR = "./outputs"
+LEVELNO_TO_LEVEL_NAME = {
+    0: "NOTSET",
+    10: "DEBUG",
+    20: "INFO",
+    30: "WARNING",
+    40: "ERROR",
+    50: "CRITICAL",
+}
 
-def get_format(name: str, level: str, distributed: bool):
-    fmt_date = '%Y-%m-%d_%T %Z'
+def rank_filter(record):
+    try:
+        return dist.get_rank() == 0
+    except RuntimeError:  # Default process group has not been initialized, please make sure to call init_process_group.
+        return True
+
+def get_format(level: str, distributed: bool = False):
     debug_and_multi_gpu = (level == 'DEBUG' and distributed)
+    fmt = "<green>{time:YYYY-MM-DD HH:mm:ss.SSS}</green> | <level>{level: <8}</level> | <cyan>{name}</cyan>:<cyan>{function}</cyan>:<cyan>{line}</cyan> - <level>{message}</level>"
+
     if debug_and_multi_gpu:
-        fmt = f'[GPU:{dist.get_rank()}] %(asctime)s | %(levelname)s\t\t| %(funcName)s:<%(filename)s>:%(lineno)s >>> %(message)s'
-    else:
-        fmt = '%(asctime)s | %(levelname)s\t\t| %(funcName)s:<%(filename)s>:%(lineno)s >>> %(message)s'
+        fmt = f"[GPU:{dist.get_rank()}] " + fmt
 
-    return fmt, fmt_date
+    only_rank_zero = not debug_and_multi_gpu
+    return fmt, only_rank_zero
 
-def add_stream_handler(distributed: bool):
-    logger = logging.getLogger(ROOT_LOGGER_NAME)
-    handler = logging.StreamHandler()
-    fmt, fmt_date = get_format(logger.name, logger.level, distributed=distributed)
+def add_stream_handler(level: str, distributed: bool):
+    fmt, only_rank_zero = get_format(level, distributed=distributed)
+    logger.add(sys.stderr, level=level, format=fmt, filter=rank_filter if only_rank_zero else "")
 
-    formatter = logging.Formatter(fmt, fmt_date)
-    handler.setFormatter(formatter)
-    if distributed:
-        handler.addFilter(RankFilter())
-
-    logger.addHandler(handler)
 
 
 def add_file_handler(log_filepath: str, distributed: bool):
-    logger = logging.getLogger(ROOT_LOGGER_NAME)
-    handler = logging.FileHandler(log_filepath)
-
-    fmt, fmt_date = get_format(logger.name, logger.level, distributed=distributed)
-    formatter = logging.Formatter(fmt, fmt_date)
-    handler.setFormatter(formatter)
-    if distributed:
-        handler.addFilter(RankFilter())
-
-    logger.addHandler(handler)
+    level = LEVELNO_TO_LEVEL_NAME[logger._core.min_level]
+    fmt, only_rank_zero = get_format(level, distributed=distributed)
+    logger.add(log_filepath, level=level, format=fmt, filter=rank_filter if only_rank_zero else "", enqueue=True)
 
 
 
-def _custom_logger(name: str, level: str, distributed: bool):
-    logger = logging.getLogger(name)
-
-    logger.setLevel(logging.INFO)
-    if not logger.hasHandlers():
-        add_stream_handler(distributed)
+def _custom_logger(level: str, distributed: bool):
+    logger.remove()
+    add_stream_handler(level, distributed)
 
     return logger
 
@@ -70,19 +65,8 @@ def set_logger(level: str = 'INFO', distributed=False):
         print(e)
         print("Skipping timezone setting.")
     _level: Literal['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL'] = level.upper()
-    _custom_logger(ROOT_LOGGER_NAME, _level, distributed)
+    _custom_logger(_level, distributed)
 
-    logger = logging.getLogger(ROOT_LOGGER_NAME)
-    if _level == 'DEBUG':
-        logger.setLevel(logging.DEBUG)
-    elif _level == 'INFO':
-        logger.setLevel(logging.INFO)
-    elif _level == 'WARNING':
-        logger.setLevel(logging.WARNING)
-    elif _level == 'ERROR':
-        logger.setLevel(logging.ERROR)
-    elif _level == 'CRITICAL':
-        logger.setLevel(logging.CRITICAL)
     return logger
 
 
@@ -104,6 +88,48 @@ def yaml_for_logging(config: DictConfig):
     config_summarized = OmegaConf.create(_yaml_for_logging(config))
     return OmegaConf.to_yaml(config_summarized)
 
+
+def _new_logging_dir(output_root_dir, project_id):
+    version_idx = 0
+    project_dir: Path = Path(output_root_dir) / project_id
+
+    while (project_dir / f"version_{version_idx}").exists():
+        version_idx += 1
+
+    new_logging_dir: Path = project_dir / f"version_{version_idx}"
+    new_logging_dir.mkdir(exist_ok=True, parents=True)
+    return new_logging_dir
+
+def _find_logging_dir(output_root_dir, project_id):
+    version_idx = 0
+    project_dir: Path = Path(output_root_dir) / project_id
+
+    while (project_dir / f"version_{version_idx + 1}").exists():
+        version_idx += 1
+
+    logging_dir: Path = project_dir / f"version_{version_idx}"
+    return logging_dir
+
+def get_logging_dir(task: str, model: str, project_id: Optional[str] = None, output_root_dir: str = OUTPUT_ROOT_DIR, distributed: bool = False) -> Path:
+    project_id = project_id if project_id is not None else f"{task}_{model}"
+
+    if not distributed:
+        return _new_logging_dir(output_root_dir, project_id)
+
+    # TODO: Better synchronization
+    if dist.get_rank() == 0:
+        logging_dir = _new_logging_dir(output_root_dir, project_id)
+        signal = torch.tensor([1]).to("cuda")
+        for rank_idx in range(1, dist.get_world_size()):
+            dist.send(tensor=signal, dst=rank_idx)
+    else:
+        signal = torch.tensor([0]).to("cuda")
+        dist.recv(tensor=signal, src=0)
+
+        logging_dir = _find_logging_dir(output_root_dir, project_id)
+
+    dist.barrier()
+    return logging_dir
 
 if __name__ == '__main__':
     set_logger(level='DEBUG')